+++ src/SpeciesDistributionToolkit.jl
@@ -24,6 +24,7 @@ import GeoJSON
 import PolygonOps
 import ZipArchives
 import Downloads
+import HTTP
 
 # Functions to get latitudes/longitudes
 include("latlon.jl")
@@ -57,9 +58,13 @@ export pseudoabsencemask, backgroundpoints
 include("polygons/polygons.jl")
 include("polygons/zonal.jl")
 include("polygons/mosaic.jl")
-include("polygons/gadm.jl")
+include("polygons/simplify.jl")
 export trim
 export zone, byzone
+
+# Boundaries
+include("boundaries/gadm.jl")
+include("boundaries/osm.jl")
 
 # Extra functions
 include("utilities.jl")