+++ docs/src/reference/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## `v1.4.2`
+
+- **added** an `openstreetmap` method (not exported) to get various GeoJSON limits from plan-text queries
+- **added** a `simplify` and `simplify!` method to remove some complexity from OSM polygons
+- **fixed** the vignettes so they don't use GADM
+
 ## `v1.4.1`
 
 - **fixed** a bug where GeoJSON GADM files were not readable by `ZipFiles`